django_visionary/requirements.txt
django
django-rest-framework
tensorflow
keras
pillow
pandas
numpy
